+++ dataquality-statistics/src/test/java/org/talend/dataquality/statistics/type/AvroDataTypeDiscoveryAnalyzerTest.java
@@ -324,12 +324,24 @@ public class AvroDataTypeDiscoveryAnalyzerTest {
             String path = AvroDataTypeDiscoveryAnalyzerTest.class.getResource("../sample/date.avro").getPath();
             Pair<Stream<IndexedRecord>, Schema> pair = streamAvroFile(new File(path));
             analyzer.init(pair.getRight());
-            analyzer.analyze(pair.getLeft()).collect(Collectors.toList());
+            pair.getLeft().forEach(analyzer::analyze);
             Schema result = analyzer.getResult();
-            // All field with "logicalType" should be detected as DATE.
-            // LogicalTypes.Date, LogicalTypes.TimeMicros, LogicalTypes.TimeMillis, LogicalTypes.TimestampMicros, LogicalTypes.TimestampMillis
-            Schema expectedSchema = new Schema.Parser().parse(
-                    "{\"type\": \"record\",\"name\": \"dateTimeLogical\",\"namespace\": \"org.talend\",\"fields\": [{\"name\": \"name\",\"type\": {\"type\": \"string\",\"talend.component.dqType\": {\"matchings\": [{\"total\": 3,\"dataType\": \"STRING\"}],\"dataType\": \"STRING\"}}},{\"name\": \"date\",\"type\": {\"type\": \"int\",\"logicalType\": \"date\",\"talend.component.dqType\": {\"matchings\": [{\"total\": 3,\"dataType\": \"DATE\"}],\"dataType\": \"DATE\"}}},{\"name\": \"timemillis\",\"type\": {\"type\": \"int\",\"logicalType\": \"time-millis\",\"talend.component.dqType\": {\"matchings\": [{\"total\": 3,\"dataType\": \"DATE\"}],\"dataType\": \"DATE\"}}},{\"name\": \"timemicros\",\"type\": {\"type\": \"long\",\"logicalType\": \"time-micros\",\"talend.component.dqType\": {\"matchings\": [{\"total\": 3,\"dataType\": \"DATE\"}],\"dataType\": \"DATE\"}}},{\"name\": \"timestampmillis\",\"type\": {\"type\": \"long\",\"logicalType\": \"timestamp-millis\",\"talend.component.dqType\": {\"matchings\": [{\"total\": 3,\"dataType\": \"DATE\"}],\"dataType\": \"DATE\"}}},{\"name\": \"timestampmicros\",\"type\": {\"type\": \"long\",\"logicalType\": \"timestamp-micros\",\"talend.component.dqType\": {\"matchings\": [{\"total\": 3,\"dataType\": \"DATE\"}],\"dataType\": \"DATE\"}}}]}");
+            // Only fields with "logicalType" equals to Date, TimestampMicros, and TimestampMillis should be detected as DATE.
+            // The two others (LogicalTypes.TimeMicros and LogicalTypes.TimeMillis) should not.
+            Schema expectedSchema = new Schema.Parser()
+                    .parse("{\"type\": \"record\",\"name\": \"dateTimeLogical\",\"namespace\": \"org.talend\",\"fields\": "
+                            + "[{\"name\": \"name\",\"type\": {\"type\": \"string\",\"talend.component.dqType\": "
+                            + "{\"matchings\": [{\"total\": 3,\"dataType\": \"STRING\"}],\"dataType\": \"STRING\"}}},"
+                            + "{\"name\": \"date\",\"type\": {\"type\": \"int\",\"logicalType\": \"date\",\"talend.component.dqType\": "
+                            + "{\"matchings\": [{\"total\": 3,\"dataType\": \"DATE\"}],\"dataType\": \"DATE\"}}},"
+                            + "{\"name\": \"timemillis\",\"type\": {\"type\": \"int\",\"logicalType\": \"time-millis\","
+                            + "\"talend.component.dqType\": {\"matchings\": [{\"total\": 3,\"dataType\": \"INTEGER\"}],\"dataType\": \"INTEGER\"}}},"
+                            + "{\"name\": \"timemicros\",\"type\": {\"type\": \"long\",\"logicalType\": \"time-micros\",\"talend.component.dqType\": "
+                            + "{\"matchings\": [{\"total\": 3,\"dataType\": \"INTEGER\"}],\"dataType\": \"INTEGER\"}}},"
+                            + "{\"name\": \"timestampmillis\",\"type\": {\"type\": \"long\",\"logicalType\": \"timestamp-millis\","
+                            + "\"talend.component.dqType\": {\"matchings\": [{\"total\": 3,\"dataType\": \"DATE\"}],\"dataType\": \"DATE\"}}},"
+                            + "{\"name\": \"timestampmicros\",\"type\": {\"type\": \"long\",\"logicalType\": \"timestamp-micros\",\"talend.component.dqType\": "
+                            + "{\"matchings\": [{\"total\": 3,\"dataType\": \"DATE\"}],\"dataType\": \"DATE\"}}}]}");
             assertEquals(expectedSchema.toString(), result.toString());
         } catch (IOException e) {
             e.printStackTrace();